+++ chopper_generator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.0.0
+
+- Multipart request
+- form url encoded
+- add jsonAPI and formUrlEncodedApi boolean to ChopperClient
+- json and formUrlEncoding are now builtin
+- `onError`, `onResponse`, `onRequest` stream
+- error converter
+
 ## 0.1.0
 
 - update dart sdk